rewrite: Android.mk
This rewrite tries to accomplish the following:

- auto-detect everything (package name, platform), so you don't need to
  modify any files.

- produce a chicken LOCAL_MODULE for use as LOCAL_SHARED_LIBRARIES
  dependencies.

- build a cross-chicken as before, but integrate as part of the
  `ndk-build` command (good idea?)

Also, it shouldn't be difficult to have it copy all built eggs into the
libs/armeabi project root.

Inspired by this project, I created an Android.mk file that integrated
Chicken into our aosp-build tree for another project. I decided to
reintegrate that work back in here, which is why this commit is so
large.